reorganized root dir, updated level orb UX
