+++ core/conf.lua
@@ -1,0 +1,8 @@
+function love.conf(t)
+  t.version = "11.3"
+  t.window.width = 960
+  t.window.height = 540
+  t.window.vsync = 1
+  t.window.msaa = 0
+  t.console = true
+end
