+++ core/game_constants.lua
