+++ core/shared.lua
